requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
grpcio==1.67.1
grpcio-tools==1.67.1
protobuf==5.28.3
httpx==0.27.2
psutil==6.1.0
pytest==8.3.3
pytest-asyncio==0.24.0
tabulate==0.9.0
